Runtime: Add ACPI table builder API to dynamically generate a DSDT/SSDT, bugref:10733 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@106342 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/misc/acpi.cpp
@@ -398,7 +398,7 @@ DECLINLINE(void) rtAcpiTblAppendByte(PRTACPITBLINT pThis, uint8_t bData)
  * @param pvData                The data to append.
  * @param cbData                Size of the data in bytes.
  */
-DECLINLINE(void) rtAcpiTblAppendData(PRTACPITBLINT pThis, const void *pvData, size_t cbData)
+DECLINLINE(void) rtAcpiTblAppendData(PRTACPITBLINT pThis, const void *pvData, uint32_t cbData)
 {
     uint8_t *pb = rtAcpiTblBufEnsureSpace(pThis, cbData);
     if (pb)
